Fixed "Failed to initialize log file" on macOS
When running "python GuiMain.py" on macOS (at least on a M1 Mac on macOS 12.1 Monterey), "pathlib.Path('~')" doesn't seem to refer to the home folder, and the error "Failed to initialize log file" is shown. Replaced this with "pathlib.Path.home()", and this seems to have fixed the issue.